feat: 커리어 섹션 디자인 변경 사항 적용
